Redo: the previous attempt resolved the same content but landed as an
ordinary commit, because the batch that hit the manual conflict had already
run 'git merge --abort'. main was therefore never a parent and the add/add
conflict on the generator came straight back. This is a real merge.

Generators: taken main's hardened copies over this branch's sandbox
originals (repo-relative ROOT vs hardcoded /home/user paths; COBie
write-target exclusions read from CobieFieldMap.cs; label-referenced
mirrors; a header that is replaced rather than stacked). Ported the one
thing main lacked, this branch's --dry-run switch, plus an ASCII '->' in the
summary line: U+2192 is unencodable on a cp1252 Windows console, so the run
raised UnicodeEncodeError after the file had already been written.

Docs conflicts were append-only; both sides kept.

+++ StingTools/Clash/ClashGrouper.cs
@@ -118,7 +118,7 @@ namespace StingTools.Core.Clash
             //     Prior code built two parallel dictionaries (byPairAndA,
             //     byPairAndB) then concatenated them — twice the dictionary
             //     work, twice the bucket allocations. One dict, one pass.
-            var byPairSideEid = new Dictionary<(string Pair, string Side, int Eid), List<ClashRecord>>();
+            var byPairSideEid = new Dictionary<(string Pair, string Side, long Eid), List<ClashRecord>>();
             foreach (var c in clashes)
             {
                 string pair = c.MatrixPairId ?? "";
